Alterando o layout das telas e modificando constraints
