--- Movies/Movies/ListaFilmesTableViewController.swift
@@ -1,105 +1,0 @@
-//
-//  ListaFilmesTableViewController.swift
-//  Movies
-//
-//  Created by Nikolas Gianoglou Coelho on 14/02/22.
-//
-
-import UIKit
-import Alamofire
-
-class ListaFilmesTableViewController: UITableViewController {
-
-  var list: [Result] = []
-  var images: [UIImage] = []
-  
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        loadData()
-        // Uncomment the following line to preserve selection between presentations
-        // self.clearsSelectionOnViewWillAppear = false
-
-        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
-        // self.navigationItem.rightBarButtonItem = self.editButtonItem
-    }
-
-    // MARK: - Table view data source
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-      return list.count
-    }
-  
-  
-  override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-    let controller = segue.destination as! DetalhesViewController
-    let selectedRow = list[tableView.indexPathForSelectedRow!.row]
-    controller.movie = selectedRow
-  }
-  
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath) as! PosterTableViewCell
-        let movie = list[indexPath.row]
-        cell.preencheViews(result: movie)
-        return cell
-    }
-  
-    func loadData(){
-      API.loadMovies { (info) in
-        if let info = info{
-          self.list += info.results
-          DispatchQueue.main.async {
-            self.tableView.reloadData()
-          }
-        }
-      }
-    }
-    
-  
-  
-    /*
-    // Override to support conditional editing of the table view.
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the specified item to be editable.
-        return true
-    }
-    */
-
-    /*
-    // Override to support editing the table view.
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            // Delete the row from the data source
-            tableView.deleteRows(at: [indexPath], with: .fade)
-        } else if editingStyle == .insert {
-            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-        }    
-    }
-    */
-
-    /*
-    // Override to support rearranging the table view.
-    override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
-
-    }
-    */
-
-    /*
-    // Override to support conditional rearranging of the table view.
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the item to be re-orderable.
-        return true
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
